extract release notes filling into separate script
returned old stage order too
